+++ src/main/java/acme/features/administrator/banner/AdministratorBannerListService.java
@@ -1,0 +1,54 @@
+
+package acme.features.administrator.banner;
+
+import java.util.Collection;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import acme.entities.offer.Banner;
+import acme.framework.components.accounts.Administrator;
+import acme.framework.components.models.Tuple;
+import acme.framework.services.AbstractService;
+
+@Service
+public class AdministratorBannerListService extends AbstractService<Administrator, Banner> {
+
+	//Internal State
+
+	@Autowired
+	protected AdministratorBannerRepository repository;
+
+	//AbstractService Interface
+
+
+	@Override
+	public void check() {
+		super.getResponse().setChecked(true);
+	}
+
+	@Override
+	public void authorise() {
+		super.getResponse().setChecked(true);
+	}
+
+	@Override
+	public void load() {
+		final Collection<Banner> banners;
+
+		banners = this.repository.findAllBanners();
+
+		super.getBuffer().setData(banners);
+	}
+
+	@Override
+	public void unbind(final Banner object) {
+		assert object != null;
+
+		Tuple tuple;
+
+		tuple = super.unbind(object, "lastModified", "displayStart", "displayFinish", "slogan", "picture", "target");
+
+		super.getResponse().setData(tuple);
+	}
+}
