+++ src/main/java/acme/features/administrator/banner/AdministratorBannerRepository.java
@@ -1,5 +1,5 @@
 
-package acme.features.any.banner;
+package acme.features.administrator.banner;
 
 import java.util.Collection;
 
@@ -10,11 +10,11 @@ import acme.entities.offer.Banner;
 import acme.framework.repositories.AbstractRepository;
 
 @Repository
-public interface AnyBannerRepository extends AbstractRepository {
+public interface AdministratorBannerRepository extends AbstractRepository {
 
 	@Query("SELECT b FROM Banner b WHERE b.id = :id")
 	Banner findOneBannerById(int id);
 
 	@Query("SELECT b FROM Banner b")
-	Collection<Banner> findAllBanners(int id);
+	Collection<Banner> findAllBanners();
 }
