+++ src/main/webapp/WEB-INF/resources/sample-data/banner.csv
@@ -1,3 +1,6 @@
 key,lastModified,displayStart,displayFinish,slogan,target,picture
 banner-01,2017/01/01 22:00,2017/01/01 22:00,2030/01/01 22:00,"slogan","https://www.youtube.com/","https://i.imgur.com/2bvab7y.jpeg"
 banner-02,2017/01/01 22:00,2017/01/01 22:00,2024/01/01 22:00,".","https://www.youtube.com/","https://i.imgur.com/aTco5L6.jpeg"
+banner-03,2022/07/30 00:00,2022/07/30 00:00,2026/03/04 11:00,"this promotion could interest you, you can check out the link and visit it!","https://www.youtube.com/","https://i.imgur.com/f1yeHVO.jpeg"
+banner-04,2022/07/30 00:00,2022/08/01 22:00,2024/01/01 22:00,"this promotion will be available next month!","https://www.youtube.com/","https://i.imgur.com/aTco5L6.jpeg"
+banner-05,2017/01/01 22:00,2017/01/01 22:00,2017/02/01 22:00,"this banner is out of its display period","https://www.youtube.com/","https://i.imgur.com/aTco5L6.jpeg"
