+++ src/main/webapp/WEB-INF/views/administrator/banner/form.jsp
@@ -1,0 +1,24 @@
+<%--
+- form.jsp
+-
+- Copyright (C) 2012-2023 Rafael Corchuelo.
+-
+- In keeping with the traditional purpose of furthering education and research, it is
+- the policy of the copyright owner to permit non-commercial use and redistribution of
+- this software. It has been tested carefully, but it is not guaranteed for any particular
+- purposes.  The copyright owner does not offer any warranties or representations, nor do
+- they accept any liabilities with respect to them.
+--%>
+
+<%@page language="java"%>
+
+<%@taglib prefix="jstl" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@taglib prefix="acme" uri="http://www.the-acme-framework.org/"%>
+
+<acme:form>
+	<acme:input-textbox code="administrator.banner.form.label.slogan" path="slogan"/>
+	<acme:input-moment code="administrator.banner.form.label.lastModified" path="lastModified"/>
+	<acme:input-moment code="administrator.banner.form.label.displayStart" path="displayStart"/>
+	<acme:input-moment code="administrator.banner.form.label.displayFinish" path="displayFinish"/>
+	<acme:input-url code="administrator.banner.form.label.target" path="target"/>
+</acme:form>
