+++ src/main/webapp/WEB-INF/views/administrator/banner/list.jsp
@@ -15,7 +15,8 @@
 <%@taglib prefix="jstl" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@taglib prefix="acme" uri="http://www.the-acme-framework.org/"%>
 
-<acme:form>
-	<acme:input-textbox code="any.banner.form.label.slogan" path="company"/>
-	<acme:input-textbox code="any.banner.form.label.moreInfo" path="moreInfo"/>
-</acme:form>
+<acme:list>
+	<acme:list-column code="administrator.banner.list.label.slogan" path="slogan"/>
+	<acme:list-column code="administrator.banner.list.label.lastModified" path="lastModified"/>
+	<acme:list-column code="administrator.banner.list.label.target" path="target"/>
+</acme:list>
