+++ src/main/webapp/WEB-INF/views/administrator/banner/messsages-en.i18n
@@ -1,0 +1,22 @@
+# messages-en.i18n
+#
+# Copyright (C) 2012-2023 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+administrator.banner.form.label.slogan = Slogan
+administrator.banner.form.label.lastModified = Last modified
+administrator.banner.form.label.displayStart = Display start date
+administrator.banner.form.label.displayFinish = Display finish date
+administrator.banner.form.label.target = Target link
+
+administrator.banner.list.label.slogan = Slogan
+administrator.banner.list.label.lastModified = Last modified
+administrator.banner.list.label.target = Target link
+
+administrator.banner.form.button.create = Register
+administrator.banner.form.button.update = Update
