+++ src/main/webapp/WEB-INF/views/administrator/banner/messsages-es.i18n
@@ -1,0 +1,22 @@
+# messages-es.i18n
+#
+# Copyright (C) 2012-2023 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+administrator.banner.form.label.slogan = Eslogan
+administrator.banner.form.label.lastModified = Última modificación
+administrator.banner.form.label.displayStart = Inicio de despliegue
+administrator.banner.form.label.displayFinish = Final de despliegue
+administrator.banner.form.label.target = Enlace
+
+administrator.banner.list.label.slogan = Eslogan
+administrator.banner.list.label.lastModified = Última modificación
+administrator.banner.list.label.target = Enlace
+
+administrator.banner.form.button.create = Registrar
+administrator.banner.form.button.update = Actualizar
