--- src/main/webapp/WEB-INF/views/any/banner/messsages-en.i18n
@@ -1,15 +1,0 @@
-# messages-en.i18n
-#
-# Copyright (C) 2012-2023 Rafael Corchuelo.
-#
-# In keeping with the traditional purpose of furthering education and research, it is
-# the policy of the copyright owner to permit non-commercial use and redistribution of
-# this software. It has been tested carefully, but it is not guaranteed for any particular
-# purposes.  The copyright owner does not offer any warranties or representations, nor do
-# they accept any liabilities with respect to them.
-
-authenticated.consumer.form.label.company = Company
-authenticated.consumer.form.label.sector = Sector
-
-authenticated.consumer.form.button.create = Register
-authenticated.consumer.form.button.update = Update
