+++ src/main/webapp/WEB-INF/views/fragments/menu-en.i18n
@@ -32,6 +32,8 @@ master.menu.administrator.user-accounts = User accounts
 master.menu.administrator.populate-initial = Populate DB (initial)
 master.menu.administrator.populate-sample = Populate DB (samples)
 master.menu.administrator.shut-down = Shut down
+master.menu.banner = Banners
+master.menu.banner.list = Banner list
 
 # master.menu.provider ########################################################
