+++ src/main/webapp/WEB-INF/views/fragments/menu-es.i18n
@@ -32,6 +32,8 @@ master.menu.administrator.user-accounts = Cuentas de usuario
 master.menu.administrator.populate-initial = Poblar BD (inicial)
 master.menu.administrator.populate-sample = Poblar BD (ejemplos)
 master.menu.administrator.shut-down = Parar el sistema
+master.menu.banner = Anuncios
+master.menu.banner.list = Lista de anuncios
 
 # master.menu.provider ########################################################
